Moved MockInstantiatorV18 to kt10
