--- karibu-testing-v10/kt10-testrun-vaadin18/src/test/java/com/vaadin/flow/component/littemplate/MyLitTemplateParserImpl.java
@@ -1,7 +1,0 @@
-package com.vaadin.flow.component.littemplate;
-
-/**
- * Workaround for https://github.com/vaadin/flow/issues/9169
- */
-public class MyLitTemplateParserImpl extends LitTemplateParserImpl {
-}
