+++ karibu-testing-v10/kt10-testrun-vaadin18/src/test/kotlin/AllTests.kt
@@ -1,15 +1,10 @@
 import com.github.mvysny.dynatest.DynaTest
 import com.github.mvysny.kaributesting.v10.*
 import com.github.mvysny.kaributesting.v10.mock.MockVaadinHelper
-import com.vaadin.flow.component.littemplate.MockInstantiatorV18
 import java.net.URL
 import kotlin.test.expect
 
 class AllTests : DynaTest({
-    beforeEach {
-        MockVaadinHelper.instantiatorFactory = { MockInstantiatorV18(it) }
-    }
-
     test("flow-build-info.json exists") {
         val res: URL? = Thread.currentThread().contextClassLoader.getResource("META-INF/VAADIN/config/flow-build-info.json")
         expect(true, "flow-build-info.json is not on the classpath!") { res != null }
