+++ karibu-testing-v10/src/main/kotlin/com/github/mvysny/kaributesting/v10/mock/MockService.kt
@@ -24,5 +24,5 @@ public open class MockService(servlet: VaadinServlet,
     override fun isAtmosphereAvailable(): Boolean = false
     override fun getMainDivId(session: VaadinSession?, request: VaadinRequest?): String = "ROOT-1"
     override fun createVaadinSession(request: VaadinRequest): VaadinSession = MockVaadinSession(this, uiFactory)
-    override fun getInstantiator(): Instantiator = MockVaadinHelper.instantiatorFactory(super.getInstantiator())
+    override fun getInstantiator(): Instantiator = MockInstantiator.create(super.getInstantiator())
 }
