+++ karibu-testing-v10/src/main/kotlin/com/github/mvysny/kaributesting/v10/mock/MockVaadinHelper.kt
@@ -18,6 +18,4 @@ public object MockVaadinHelper {
             servlet.servletContext.setInitParameter("vaadin.frontend.token.file", tokenFile.absolutePath)
         }
     }
-
-    public var instantiatorFactory: (Instantiator) -> Instantiator = { MockInstantiator(it) }
 }
